requirements.txt
Flask==3.1.2
flasgger==0.9.7.1
PyJWT==2.10.1
pandas==2.3.3
requests==2.32.5
numpy==2.2.6
bs4==0.0.2
gunicorn==22.0.0
python-dotenv==1.0.1